Django-App/order_service/serializers.py
from rest_framework import serializers
from .models import Product, StockLevel, Order


class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = "__all__"


class StockLevelSerializer(serializers.ModelSerializer):
    class Meta:
        model = StockLevel
        fields = "__all__"
        depth = 2
