Django-App/requirements.txt
appdirs==1.4.3
attrs==18.2.0
black==18.9b0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Django==2.1.2
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.7
pipenv==2018.7.1
psycopg2==2.7.6.1
pytz==2018.5
requests==2.20.1
toml==0.10.0
urllib3==1.24.1
virtualenv==16.0.0
virtualenv-clone==0.3.0
